[MOD] Programmers Lv1.

+++ Programmers/Level1/부족한금액계산하기.java
@@ -1,6 +1,6 @@
 class Solution {
     public long solution(int price, int money, int count) {
-    	int result = 0;
+    	long result = 0;
     	for(int i=1; i<=count; i++) {
     		result += price * i;
     	}
